re-organize many utilities
