+++ tests/utils/test_typeutils.py
@@ -8,8 +8,7 @@ from mock import Mock
 
 from tfsnippet.stochastic import StochasticTensor
 from tfsnippet.utils import (is_integer, is_float, TensorWrapper,
-                             is_tensor_object, TensorArgValidator,
-                             convert_to_tensor_and_cast)
+                             is_tensor_object, TensorArgValidator)
 
 if six.PY2:
     LONG_MAX = long(1) << 63 - long(1)
@@ -78,26 +77,6 @@ class IsTensorObjectTestCase(unittest.TestCase):
                 msg='{!r} should not be interpreted as a tensor object'.
                     format(obj)
             )
-
-
-class ConvertToTensorTestCase(tf.test.TestCase):
-
-    def test_convert_to_tensor(self):
-        def check(x, dtype=None):
-            z = tf.convert_to_tensor(x)
-            y = convert_to_tensor_and_cast(x, dtype)
-            self.assertIsInstance(y, tf.Tensor)
-            if dtype is not None:
-                self.assertEqual(y.dtype, dtype)
-            else:
-                self.assertEqual(y.dtype, z.dtype)
-
-        check(np.arange(10, dtype=np.float32))
-        check(np.arange(10, dtype=np.float32), np.float64)
-        check(np.arange(10, dtype=np.float32), tf.float64)
-        check(tf.range(10, dtype=tf.float32))
-        check(tf.range(10, dtype=tf.float32), np.float64)
-        check(tf.range(10, dtype=tf.float32), tf.float64)
 
 
 class TensorArgValidatorTestCase(tf.test.TestCase):
