+++ tfsnippet/ops/__init__.py
@@ -4,12 +4,18 @@ from .control_flows import *
 from .convolution import *
 from .evaluation import *
 from .misc import *
+from .shape_utils import *
 from .shifting import *
+from .type_utils import *
 
 __all__ = [
     'add_n_broadcast', 'assert_rank', 'assert_rank_at_least',
     'assert_scalar_equal', 'assert_shape_equal', 'bits_per_dimension',
-    'classification_accuracy', 'depth_to_space', 'log_mean_exp', 'log_sum_exp',
-    'maybe_clip_value', 'shift', 'smart_cond', 'softmax_classification_output',
-    'space_to_depth',
+    'broadcast_concat', 'broadcast_to_shape', 'broadcast_to_shape_strict',
+    'classification_accuracy', 'convert_to_tensor_and_cast', 'depth_to_space',
+    'flatten_to_ndims', 'log_mean_exp', 'log_sum_exp', 'maybe_clip_value',
+    'prepend_dims', 'reshape_tail', 'shift', 'smart_cond',
+    'softmax_classification_output', 'space_to_depth', 'transpose_conv2d_axis',
+    'transpose_conv2d_channels_last_to_x',
+    'transpose_conv2d_channels_x_to_last', 'unflatten_from_ndims',
 ]
