+++ tfsnippet/ops/convolution.py
@@ -1,7 +1,7 @@
 import tensorflow as tf
 
-from tfsnippet.utils import (add_name_arg_doc, flatten_to_ndims,
-                             unflatten_from_ndims)
+from tfsnippet.utils import add_name_arg_doc
+from .shape_utils import flatten_to_ndims, unflatten_from_ndims
 
 __all__ = ['space_to_depth', 'depth_to_space']
