+++ tfsnippet/utils/__init__.py
@@ -34,32 +34,28 @@ __all__ = [
     'IntConfigValidator', 'InvertibleMatrix', 'NoReentrantContext',
     'ParamSpec', 'PermutationMatrix', 'RarExtractor', 'StatisticsCollector',
     'StrConfigValidator', 'SummaryCollector', 'TFSnippetConfig',
-    'TarExtractor', 'TemporaryDirectory', 'TensorArgValidator',
+    'TarExtractor', 'TemporaryDirectory', 'TensorArgValidator', 'TensorSpec',
     'TensorWrapper', 'VarScopeObject', 'VarScopeRandomState', 'ZipExtractor',
     'add_histogram', 'add_name_and_scope_arg_doc', 'add_name_arg_doc',
     'add_summary', 'append_arg_to_doc', 'append_to_doc', 'assert_deps',
-    'broadcast_concat', 'broadcast_to_shape', 'broadcast_to_shape_strict',
-    'camel_to_underscore', 'concat_shapes', 'convert_to_tensor_and_cast',
-    'create_session', 'default_summary_collector', 'deprecated',
-    'deprecated_arg', 'ensure_variables_initialized', 'flatten_to_ndims',
-    'generate_random_seed', 'get_batch_size', 'get_cache_root',
-    'get_config_defaults', 'get_config_validator', 'get_default_scope_name',
-    'get_default_session_or_error', 'get_dimension_size',
-    'get_dimensions_size', 'get_model_variables', 'get_rank',
-    'get_reuse_stack_top', 'get_shape', 'get_static_shape',
+    'camel_to_underscore', 'concat_shapes', 'create_session',
+    'default_summary_collector', 'deprecated', 'deprecated_arg',
+    'ensure_variables_initialized', 'generate_random_seed', 'get_batch_size',
+    'get_cache_root', 'get_config_defaults', 'get_config_validator',
+    'get_default_scope_name', 'get_default_session_or_error',
+    'get_dimension_size', 'get_dimensions_size', 'get_model_variables',
+    'get_rank', 'get_reuse_stack_top', 'get_shape', 'get_static_shape',
     'get_uninitialized_variables', 'get_variable_ddi', 'get_variables_as_dict',
     'global_reuse', 'humanize_duration', 'instance_reuse', 'is_float',
     'is_integer', 'is_shape_equal', 'is_tensor_object',
     'is_tensorflow_version_higher_or_equal', 'iter_files', 'makedirs',
     'maybe_add_histogram', 'maybe_check_numerics', 'maybe_close',
-    'minibatch_slices_iterator', 'model_variable', 'prepend_dims',
-    'print_as_table', 'register_config_arguments', 'register_config_validator',
-    'register_tensor_wrapper_class', 'reopen_variable_scope', 'reshape_tail',
+    'minibatch_slices_iterator', 'model_variable', 'print_as_table',
+    'register_config_arguments', 'register_config_validator',
+    'register_tensor_wrapper_class', 'reopen_variable_scope',
     'resolve_negative_axis', 'root_variable_scope', 'scoped_set_config',
     'set_cache_root', 'set_random_seed', 'settings', 'split_numpy_array',
-    'split_numpy_arrays', 'transpose_conv2d_axis',
-    'transpose_conv2d_channels_last_to_x',
-    'transpose_conv2d_channels_x_to_last', 'unflatten_from_ndims',
-    'validate_enum_arg', 'validate_group_ndims_arg', 'validate_int_tuple_arg',
-    'validate_n_samples_arg', 'validate_positive_int_arg',
+    'split_numpy_arrays', 'validate_enum_arg', 'validate_group_ndims_arg',
+    'validate_int_tuple_arg', 'validate_n_samples_arg',
+    'validate_positive_int_arg',
 ]
